refactor: package 이동
state package -> game/state package
board/exception package -> game/exception package
