+++ src/main/java/chess/ConsoleChessApplication.java
@@ -2,7 +2,7 @@ package chess;
 
 import chess.controller.ChessGameController;
 import chess.domain.game.ChessGame;
-import chess.domain.state.Ready;
+import chess.domain.game.state.Ready;
 
 public class ConsoleChessApplication {
 	public static void main(String[] args) {
