+++ src/test/java/chess/domain/game/ChessGameTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import chess.domain.state.Ready;
+import chess.domain.game.state.Ready;
 
 public class ChessGameTest {
 	@Test
